Started basic visualization methods. First simple methods successfuly implemented and tested on synthetic data. Implemented two methods as common flow and comet visualization. Rendering time is OK, around 0.123us. So far no customization from GUI is implemented. Grid rendering is done and separated to three grid levels.

+++ DACUSN/stackmanager.cpp
@@ -234,7 +234,7 @@ void stackManager::dataProcessing(rawData *data)
     {
         for(i = 0; i<radarList->count(); i++)
         {
-            if(radarList->at(i)->id==radar_id) break;
+            if(radarList->at(i)->id==(unsigned int)(radar_id)) break;
         }
     }
